changed float to double (#8, #6)

+++ client/Research2DPrototype/Assets/Scripts/ElementData.cs
@@ -6,7 +6,7 @@ order = 51)]
 public class ElementData : ScriptableObject
 {
   [SerializeField]
-  private float Value;
+  private double Value;
   [SerializeField]
   private int Power;
   [SerializeField]
